requirements.txt
paho-mqtt
python-dotenv
openai
langchain
structlog
langchain-community
pytest
langchain-ollama
